Creates VI for IIR filtering data, adds a VI for getting RPM of shooter wheels, makes VI to detect when RPM is in range, Begins work on Auto Shoot Mode.
TODO:
Make warmup wait until the RPM is ready, Angle is ready, and there is a ball.

Make the shooter not shoot when the ball has just been loaded. Add a dashboard control that adds a time delay.

Integrate with Intake subsystem to communicate what should be done with the trigger